+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 ## 1.0.5 (2019-03-16)
+ * Added robots update form in modal
  * Added dashboard navigation items for NavBar
  * Added visitor details modal
  * Added bot detection